Break tie in FillerTest.HugePageFrac.
When we release a1 and a6, we have two huge pages with an equal number of huge
pages.  The order that we allocate the huge pages effectively implicitly breaks
the tie when iterating with loop() in HugePageFiller::ReleasePages.

Since we are concerned with the hugepage_frac() calculation, modify the test so
that our choices are unambiguous under our current subrelease policy.

PiperOrigin-RevId: 305468484
Change-Id: I5bfbc2d5182fb46f69b9aed394234d8d274621fb