Optimize HugePageFiller::ReleasePages for releasing up to a huge page.
In ad2aca145fcd, we plumbed the total number of pages that we are trying to
subrelease to the HugePageFiller.  Rather than iterate through all huge pages
to select a best candidate and start over if that is insufficient, we can use
the requested length to do fewer passes.

In the common case of needing to subrelease a small amount of memory scattered
across many pages, we only need to do a single linear pass through our
collection of huge pages in the filler.

PiperOrigin-RevId: 305877700
Change-Id: I43868ecedc70aaf0d0011a302a87ad604f8576ae