Optimize Push/Pop inlined RSEQ code of tcmalloc
The old code had a hotspot for a partial flag update / stall from:

asm (.... "decl (scratch, cl, 8)" : ...)
(c++ code)
test overflow
jae ....

This code removes this by using LEA for decreasing the 'current' value. This has the additional effect that flags are no longer affected in the submit portion of the code, which means we can directly export the AE (or BE) flags as overflow/underflow.

The compiler then DTRT by turning the "if (underflow)" into a direct "jae ...." construct.

This is applied for both pop (underflow) and (push) overflow

PiperOrigin-RevId: 312185620
Change-Id: I121376bf295417fd8af8947278071ee105d826fd