Better TCMalloc Release Decisions.
If the current amount of allocated pages (free or used) is below a recent peak of demand for pages, do not break up any further hugepages. "Recent" is a command line configurable parameter. Also tracks the fraction of these decisions that were correct (i.e., soon after choosing not to release memory, demand exceeded the amount of allocated memory at the time we made the decision). At the moment, "soon" is the same as "recent".

PiperOrigin-RevId: 314236718
Change-Id: I518c0c923514b118d34f6f0099f9f080c3f999e9