Implement virtual CPU support and provide rseq ABI extension.
This is a prototype extension to the rseq() syscall.  Since a process may run on only a few cores at a time, we can use a dense set of "v(irtual) cpus."  This can reduce cache requirements, as we only need N caches for the cores we actually run on simultaneously, rather than a cache for every physical core.

This can reduce the RAM footprint of caches (reducing metadata since lazy cache initialization [24f41890cbe4], freelisted objects) and improve hit rates, since a cache is more likely to be recently used even after core migration.

PiperOrigin-RevId: 318347733
Change-Id: I91eda0c35bfd0a7b658dbee38e5c1a469191b458