Exercise wraparound behavior in LockFreeTransferCache.
We work with a batch size that is coprime relative to
std::numeric_limits<uint32_t>::max()+1.  By pushing/popping batches repeatedly,
one of the inserts will trigger a wraparound that spans both the head and tail
of the circular buffer.

This was fixed by adc469884af0, which ensures that the LockFreeTransferCache
also encounters a wraparound early (within kMaxObjectsToMove).  This test fails
prior to that commit and passes with it.

PiperOrigin-RevId: 340840534
Change-Id: Id90e461f2ba33459a0b4a1829fa2af87a4f9a97c