Implement NUMA awareness
Modern systems increasingly rely upon NUMA architectures for scalability,
grouping CPUs into a number of NUMA nodes which each have fast to access local
memory & slower access to remote memory. Up to now TCMalloc has not been
NUMA-aware, meaning that it shares memory between all NUMA nodes & makes no
attempt to ensure that allocations occur using memory local to the CPU
performing the allocation.

Here we introduce an implementation of NUMA awareness aimed at improving the
behavior of processes that run on CPUs spanning multiple NUMA nodes.

In general we cannot know at compile time how many NUMA nodes the system that
we run upon will include, but we also cannot size our data structures
arbitrarily at runtime in the name of efficiency. In order to resolve the
conflict between these two constraints we define the concept of a NUMA
'partition' as being an arbitrary set of NUMA nodes, disjoint from all other
partitions. At compile time we select a fixed number of partitions to support,
and at runtime we map each NUMA node in the system to a partition.  If the
number of supported partitions is greater than or equal to the number of NUMA
nodes in the system then partition & node are effectively identical.  If
however the system has more nodes than we do partitions then nodes assigned to
the same partition will share size classes & thus memory. This may incur a
performance hit, but allows us to at least run on any system.

We duplicate our set of size classes such that each NUMA partition has its own,
and when performing an allocation we use the size class corresponding to the
partition to which the CPU executing the allocation is assigned. We duplicate
the page allocator for each NUMA partition to prevent memory being reused for
an allocation occurring on a CPU assigned to a different partition, and the
system allocator binds mapped memory to the set of NUMA nodes contained within
the appropriate partition. The end result is that allocations should always
return local memory.

Of note is that the determination of which partition is local may change within
the context of a single thread unless that thread is affine only to CPUs that
are all local to one partition, because that thread may migrate to a CPU local
to another node & another partition. In general, this means that users ought
only to expect that allocated memory is guaranteed to be local if they
constrain the allocating thread to CPUs within a single NUMA node. Since a
thread that isn't constrained this way may migrate between allocating & using
memory anyway TCMalloc isn't imposing any extra limitations here.

Currently this NUMA awareness is not enabled anywhere - kNumaPartitions is
unconditionally 1 & this causes the implementation of NUMA awareness to
effectively be optimized away by the compiler.

PiperOrigin-RevId: 373607321
Change-Id: Ic30fdabc4899fe8814eb3d3bed73446b0625edd8