Define kNumClasses as a multiple of new kNumBaseClasses
We currently have a fixed base number of size classes for each
TCMALLOC_PAGE_SHIFT configuration & explicitly multiply that by 2 in each
declaration of kNumClasses when kHasExpandedClasses is true.

Work to introduce NUMA awareness to TCMalloc will similarly multiply the number
of size classes - the current set of non-extended size classes becomes specific
to one (set of) NUMA node(s), then we duplicate them for the next (set of) NUMA
node(s) etc and finally we have the current extended classes.

In order to simplify the addition of the extra sets of size classes here we
factor out kNumClasses into the number of non-extended classes which we
designate kNumBaseClasses and then we centralize the multiplication of that to
account for kHasExpandedClasses in one definition of kNumClasses. Various
pieces of code which explicitly used kNumClasses/2 are adjusted to instead use
kNumBaseClasses, which will help them adapt when extra classes are added for
NUMA awareness.

PiperOrigin-RevId: 378726014
Change-Id: I49be21b9f713f6a49ae4a0243481ed3418dbdd63