Allow variable TcmallocSlab CPU shifts
Currently the CPU shift used by TcmallocSlab is a compile-time constant
provided as a template argument. NUMA awareness is going to require a larger
shift value, but it's desirable to avoid requiring the corresponding increase
in per-CPU memory usage when NUMA awareness is present but disabled (e.g.
because the system doesn't have multiple NUMA nodes or because the user
requested that it be disabled).

Make the shift a variable set during Init() instead of a template argument,
allowing it to be chosen at runtime. The costs of this are:

1) An extra load of the shift value when manipulating the slab. The value is
   stored in the same cache line as slabs_ which we already have to load - i.e.
   it will certainly be a cache hit & thus pretty inexpensive.

2) We naturally can't encode the shift as an immediate anymore.

3) For aarch64 we can't use the Operand2 barrel shift & instead perform the
   shift using an extra lsl instruction.

These costs seem low enough to justify making shift unconditionally a variable
for simplicity.

Experimentally this costs ~0.1ns for BM_new_delete on my Xeon W-2135 machine.

PiperOrigin-RevId: 380212016
Change-Id: I55b6359df3de79f18a371d9b51a54340d3e0c572