tcmalloc: more thoroughly apply `HAVE_STRUCT_MALLINFO`
This more conservatively applies the `HAVE_STRUCT_MALLINFO` guards to the
`mallopt` interfaces enabling tcmalloc to be more portable for libcs which do
not expose the SVID compatible `mallopt` interfaces or `struct mallinfo`.

The known libc implementations that provide `mallopt` include:
- glibc (and variants, e.g. eglibc)
- uclibc
- bionic
- newlib

In contrast, musl is known to not provide the `mallopt` interface.

PiperOrigin-RevId: 391076097
Change-Id: I252ea1f51302e53ebafa564a609c905c5087178d