remove usage of `sys/cdefs.h`
Remove the usage of the `sys/cdefs.h` header which is a BSDism which migrated to
Linux via glibc.  This header is used to provide some defines related to the C
language.  This was primarily being used for acquiring the definition of
`__THROW`.  Excise the references to this header to allow better portability.

PiperOrigin-RevId: 391083181
Change-Id: Ieb6e26fbb7b2cae353309080f4c4f9cdd77d2b50