tcmalloc: make tcmalloc more portable to libc implementations
`TEMP_FAILURE_RETRY` is a non-standard macro to help with handling calls to
functions which may fail with `EINTR`.  This macro is not defined in all libc
implementations as it is not part of the C standard.  This adds an equivalent
macro: `TCMALLOC_RETRY_ON_TEMP_FAILURE` so that tcmalloc can be easily built
against more libc implementations.
PiperOrigin-RevId: 396843617
Change-Id: I293f609fbe5e1ddabfda6d85ea9a05fed6b5368c