Implement interfaces for providing access frequency hints to TCMalloc.
This adds an additional page heap, identified with a high-order bit to indicate
the allocation-time expected access pattern.  This spares us the need to lookup
whether or not an allocation was allocated "as hot" during deallocation
(defeating the value of C++14's sized delete feature).

The frontend encodes the access frequency hint into the size class, which we
use to decide which page heap instance should be used.  This avoids physically
duplicating most of the front-end data structures (per-CPU/transfer
cache/central freelist).  Cold size classes appear consecutively, as the code
for handling alignment relies on being able to increment the size class to move
to the next one.  Incrementing the size class by 1 should not change whether
the allocation is going to be treated as hot or cold.

All small objects are considered frequently accessed, based on empirical data
and as an implementation convienence.  For large objects, we represent free
objects without full use of Span's intrusive linked list, which would otherwise
incur memory accesses solely for managing their allocation.

PiperOrigin-RevId: 404797344
Change-Id: I3ae07d0eb3b3712bb13f04520927649302aca10e