Fix low water mark logic, and move out of start when plundering.
Before, a sequence of plunder->InsertRange->plunder would lead to the newly
inserted elements to be removed too instead of only the old elements. So it
would return more elements to the central cache than required.

When we return elements to the freelist during plundering, we should return
old elements from the start of the ringbuffer.

This only affects the shared transfer cache, as the others don't use plundering.

PiperOrigin-RevId: 411735882
Change-Id: I08ecd86bb725541bb4644b19c1a0b02538f1a03c