Implement flag-guarded dynamic slab growth functionality.
Motivation: the per-cpu slab currently starts with 256 KiB per core, and we can potentially start with a much smaller slab to save memory and grow as needed.

We use a simple algorithm to decide whether growth is necessary based on the ratio of overflows to underflows (we don't keep track of cache hits so can't use hit ratio).

PiperOrigin-RevId: 432507550
Change-Id: I7cb1a0f91424941aff09f2f172345758e3eb94b4