Fix code for possible clobbering of x16 register on ARM
The restart sequence of inlined RSEQ assembly code is explicitly put in the text.unlikely section. The restart sequence performs a direct branch to the start of the sequence. On X86 a branch (JMP) has no practical distance limit.

On ARM however, we may hit the limit of the relative B branch instruction, and given enough distance between the restart sequence in text.unlikely and the start of the sequence, the branch is adjusted into a thunk which clobbers x16 without the compiler knowing about it.

This change adds x16 as an explicit clobber to the inlined assembly for ARM

PiperOrigin-RevId: 432549922
Change-Id: I4ea9e095fa02ec82e49d66c932e644be4ebd85fb