Rolling forward the dynamic slab sizes change after fixing bugs related to NumaShift().
One bug was that in NumaShift(), we weren't checking `topology.numa_aware()`, which meant that we would sometimes incorrectly end up using a numa_shift of 1 instead of 0 in GetShiftMaxCapacity, which caused us to use only half of the slab.

Another bug was that in GrowSlabIfNeeded(), we didn't account for numa_shift when checking whether the shift had reached the max allowed value.

PiperOrigin-RevId: 434491155
Change-Id: Id7df18cacfa86193c3a6aea7b3f1fff83aa85c39