Add support for shrinking the slab (in addition to growing) when dynamic slabs are enabled.
The motivation is to prevent cases in which we have a burst of usage and grow the slab, but then later having a larger slab is unnecessary. We set the default shrink threshold to 0.5, which is significantly different from the default growth threshold of 0.9 in order to prevent thrashing back and forth between growing/shrinking.

PiperOrigin-RevId: 438610202
Change-Id: I1b8ae2b16900c4d605f76279971882a3aeacc889