In UpdateCapacity, move the initial loading of max_capacity and capacity after the call once and add a comment nearby.
Moving the initializations after the call once makes it more likely that they will be accurate since the call once can take a lock and may be blocked on ResizeSlabs, which could make the values of capacity/max_capacity outdated. Note that this isn't a correctness issue because the code should be robust against using outdated values here, but it could be helpful to use fewer outdated values.

PiperOrigin-RevId: 443118974
Change-Id: I7a2c945dfc5acb3e534f8680828cc0fa45ffa963