Consolidate the PushBatch/PopBatch RSEQ functions by passing virtual_cpu_id_offset as another parameter.
On x86-64, we need one more register to use because (a) we can't clobber the parameter registers during RSEQ since we might need to restart, and (b) r12-r15 are callee-saved - [reference](https://stackoverflow.com/questions/18024672/what-registers-are-preserved-through-a-linux-x86-64-function-call) - so we need to preserve their values if we modify them. We choose to save the value of r12 on the redzone or stack (depending on PIE).

On AARCH64, we aren't using x15, which is caller-saved - [reference](https://developer.arm.com/documentation/den0024/a/The-ABI-for-ARM-64-bit-Architecture/Register-use-in-the-AArch64-Procedure-Call-Standard/Parameters-in-general-purpose-registers) - so we can just use x15.

In percpu_tcmalloc.h, we remove some `#if TCMALLOC_PERCPU_USE_RSEQ`: this simplifies the code and turns UB into crashing in the case of these functions being called when TCMALLOC_PERCPU_USE_RSEQ is false.

Note: for x86-64, FETCH_CPU currently does a movl and FETCH_VCPU does movzwl. We use movzwl here, which is what the inline Push/Pop RSEQ asm does in percpu_tcmalloc.h.
PiperOrigin-RevId: 449003839
Change-Id: I0e05b0198c128e06d96bd4e3ea38b8b751a86619