Resize transfer caches on a background thread.
This provides a mechanism to resize transfer caches on a background thread instead of on each overflow. If enabled, a background operation attempts to resize up to 10% of the transfer caches with the highest misses. Each candidate cache is grown by a batch size by stealing capacity from other caches that suffered relatively fewer misses.

PiperOrigin-RevId: 456346364
Change-Id: Ifd6c07841d79f399f75c824201a8d07cb3b4246c