Move `SampledAllocationAllocator` in `static_vars.h` into its own header file.
Update the name of the associated unit test.

This prepares the change to make heap profiling collection free from the global lock. We are refactoring `peak_heap_tracker` so that its linked list implementation would be based on `sampled_allocation_recorder` and would depend on `SampledAllocationAllocator`. However, `SampledAllocationAllocator` is in `static_vars.h` which depends on `peak_heap_tracker`, resulting in a dependency cycle. So moving `SampledAllocationAllocator` to its own header file.

PiperOrigin-RevId: 459796725
Change-Id: Id29109051612e75217d29d8d2308324e00c262ce