Don't call MADV_REMOVE or memset to zero the old slabs before reuse.
We were doing this in order to ensure that the old slabs were guaranteed to be zeroed before reuse, but this actually isn't necessary - the all locked headers state is fine too. MADV_REMOVE was always failing because it doesn't work on anonymous memory and then we were calling memset, which forced the whole slab to become resident.

Make the StressThreadTest simulate madvise failure sometimes by not calling madvise between calls to ResizeSlabs.

PiperOrigin-RevId: 459803710
Change-Id: I34288fbee900cab02ef836db7a82236c3498ad7a